Fix unpredictable error with SED
Not safe to write and read from same file.

+++ Makefile
@@ -3,13 +3,14 @@ SCRIPT_PATH= Scripts/PseudoDashboard.js
 
 # Packs the widget, creating a copy which will be devoured by Mac OS X
 # when the widget is installed. Also removes the development library.
-pack: Counting\ Sheep.wdgt
-	cp -r Counting\ Sheep.wdgt $(PACKED)
-	cat $(PACKED)/Main.html | sed 'N;s_\n  <script type="text/javascript" src="$(SCRIPT_PATH)"></script>__' > $(PACKED)/Main.html
+pack: clean Counting\ Sheep.wdgt
+	cp -rfv Counting\ Sheep.wdgt $(PACKED)
+	cat $(PACKED)/Main.html | sed 'N;s_\n  <script type="text/javascript" src="$(SCRIPT_PATH)"></script>__' > $(PACKED)/Main.html.tmp
+	mv $(PACKED)/Main.html.tmp $(PACKED)/Main.html
 	rm $(PACKED)/$(SCRIPT_PATH)
 
 test: pack
 	open $(PACKED)
 
-clean: $(PACKED)
-	rm -r $(PACKED)
+clean:
+	if [ -d $(PACKED) ]; then rm -r $(PACKED); fi
